We cannot really have publish task in parallel. I dump attempt to automate publishing.
The thing is, all those GitHub Actions descriptions or blog posts are useless
because all of their examples only target "on: push" i.e. they don't expect
"on: pull_request". And there is no exact documentation on how to filter
the event trigger (push) to specify on `if:` on job steps.

There might be some way to achieve that but at this state I don't plan to
waste my time on it. dotnet is not my target platform these days.